Started working on removal of service route

+++ src/customHooks/useFetchDeleteRoute.js
@@ -1,4 +1,4 @@
-import { useContext, useState, useEffect } from 'react';
+import { useContext, useState } from 'react';
 import { SubscriberContext } from 'globalState/SubscriberContext';
 
 const useFetchDeleteRoute = (lineId) => {
@@ -8,35 +8,37 @@ const useFetchDeleteRoute = (lineId) => {
 
   const confirmData = { lineId, secret };
 
-  useEffect(() => {
-    // If the user is a newUser then fire off a request to confirm them and their lines
-
-    fetch(`${process.env.REACT_APP_API_HOST}api/person/${subscriberState.query.user}`, {
-      method: 'DELETE',
-      body: JSON.stringify(confirmData),
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    })
-      .then((response) => {
-        // If the response is successful(200: OK) or error with validation message(400)
-        if (response.status === 200 || response.status === 400) {
-          return response.json(); // Return response as json
-        }
-        throw new Error(response.statusText, response.Message); // Else throw error and go to our catch below
+  const removeRoute = () => {
+    if (lineId) {
+      fetch(`${process.env.REACT_APP_API_HOST}api/person/${subscriberState.query.user}`, {
+        method: 'DELETE',
+        body: JSON.stringify(confirmData),
+        headers: {
+          'Content-Type': 'application/json',
+        },
       })
-      // If formsubmission is successful
-      .then((data) => {
-        console.log({ data });
-        setIsFetching(false); // set to false as we are done fetching now
-      }) // If formsubmission errors
-      .catch((error) => {
-        // eslint-disable-next-line no-console
-        console.error({ error });
+        .then((response) => {
+          // If the response is successful(200: OK) or error with validation message(400)
+          if (response.status === 200 || response.status === 400) {
+            return response.json(); // Return response as json
+          }
+          throw new Error(response.statusText, response.Message); // Else throw error and go to our catch below
+        })
+        // If formsubmission is successful
+        .then((data) => {
+          console.log({ data });
+          setIsFetching(false); // set to false as we are done fetching now
+        }) // If formsubmission errors
+        .catch((error) => {
+          // eslint-disable-next-line no-console
+          console.error({ error });
 
-        setIsFetching(false); // set to false as we are done fetching now
-      });
-  }, [confirmData, subscriberState.query.user]);
+          setIsFetching(false); // set to false as we are done fetching now
+        });
+    }
+  };
+
+  return { removeRoute };
 };
 
 export default useFetchDeleteRoute;
